Create site controller.
Now our app.rb only does assets. Very useful.
